Added txt list of possible stats
Considering loading a hard list of possible stat formats
and comparing against each instead of trying to figure it all
out dynamically. Might be really slow, though, we'll just have
to try it out and see.

Could be a good mix of both is the answer here -- keep doing  what
we're doing for those that work, compare against some hard-coded
formats for those that do not.